fix: 설정 정리, 라우트 정리
